% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{core_school_codes}
\alias{core_school_codes}
\title{Core School Codes}
\format{
A numeric vector of school codes. Use to filter other results.
}
\usage{
core_school_codes
}
\description{
School codes for elementary, middle, and high schools in CCPS
}
\keyword{datasets}
