% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{course_codes_lookup}
\alias{course_codes_lookup}
\title{Course Code Lookup}
\format{
A tibble with 41 rows and 3 variables:
\describe{
\item{code}{course code prefix}
\item{domain}{content domain for the prefix}
\item{specialty_center_domain}{indicator of whether prefix is specific to a specialty center}
}
}
\usage{
course_codes_lookup
}
\description{
a lookup table containing code prefixes and domains for CCPS courses
}
\keyword{datasets}
