% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fm_ids.R
\name{fetch_fm_ids}
\alias{fetch_fm_ids}
\title{Fetch Fall Membership Ids}
\usage{
fetch_fm_ids(year)
}
\arguments{
\item{year}{year to get fall membership for}
}
\value{

}
\description{
Fetch Fall Membership Ids
}
\examples{
\dontrun{
fetch_fm_ids(2019)
}
}
