% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_school_demos.R
\name{fetch_school_demos}
\alias{fetch_school_demos}
\title{Fetch School Demographics}
\usage{
fetch_school_demos(year)
}
\arguments{
\item{year}{year to get demographics for}
}
\value{

}
\description{
This will get the sex, ethnicity, sped, ell, and gifted \%s by school for a given year
}
