% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_single_demo}
\alias{get_single_demo}
\title{Get Demographics for an Entity}
\usage{
get_single_demo(df, demovar, ...)
}
\arguments{
\item{df}{data}

\item{demovar}{a string. Variable to get demographics for (e.g. sex, ethnicity)}

\item{...}{grouping variable(s)}
}
\value{

}
\description{
Get demographics for a given entity or entities.
Most useful values to pass to ... include schools/school numbers,
school levels/types, grades, etc.
}
