% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_demos.R
\name{recode_demos}
\alias{recode_demos}
\title{Recode Demographics}
\usage{
recode_demos(
  df,
  speced = indicator_speced,
  ell = ell_status,
  gate = ccps_gate,
  sex = sex
)
}
\arguments{
\item{df}{data}

\item{speced}{special education indicator. Presumed to be indicator_speced.}

\item{ell}{ell indicator. Presumed to be ell_status.}

\item{gate}{gifted indicator. Presumed to be ccps_gate.}

\item{sex}{sex indicator. Presumed to be sex.}
}
\value{

}
\description{
This will recode special education status, ell status, and gifted status.
The function assumes that janitor::clean_names() has already been run.
}
\examples{
\dontrun{
library(tidyverse)
library(janitor)
library(ccpsr)

con <- set_con()

ex <- odbc::dbGetQuery(con, 'SELECT *
FROM [CCPS_Shared].[CCPS_Shared].[RE_Students]
WHERE School_Year = 2019')

ex \%>\%
clean_names() \%>\%
recode_demos()

}
}
