% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ethnicity.R
\name{recode_ethnicity}
\alias{recode_ethnicity}
\title{Recode Ethnicity}
\usage{
recode_ethnicity(df, eth_var = ethnic, hisp_var = hispanic_latino)
}
\arguments{
\item{df}{data}

\item{eth_var}{ethnicity variable. Assumed to be "ethnic" as a result of db query and clean_names}

\item{hisp_var}{hispanic/latino variable. Assumed to be "hispanic_latino" as a result of db query and clean_names}
}
\value{

}
\description{
Recode Ethnicity
}
\examples{
\dontrun{
library(tidyverse)
library(janitor)
library(ccpsr)

con <- set_con()

ex <- odbc::dbGetQuery(con, 'SELECT *
FROM [CCPS_Shared].[CCPS_Shared].[RE_Students]
WHERE School_Year = 2019')

ex \%>\%
clean_names() \%>\%
recode_ethnicity()

}
}
