% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{secondary_school_codes}
\alias{secondary_school_codes}
\title{Secondary School Codes}
\format{
A character vector of school codes. Use to filter other results.
}
\usage{
secondary_school_codes
}
\description{
School codes for middle, high, alternative, and regional schools in CCPS
}
\keyword{datasets}
